+++ pkg/quarterdeck/tokens/claims.go
@@ -3,7 +3,7 @@ package tokens
 import (
 	jwt "github.com/golang-jwt/jwt/v4"
 	"github.com/oklog/ulid/v2"
-	ulids "github.com/rotationalio/ensign/pkg/utils/ulid"
+	"github.com/rotationalio/ensign/pkg/utils/ulids"
 )
 
 // Claims implements custom claims for the Quarterdeck application.
