+++ pkg/tenant/db/tokens_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/rotationalio/ensign/pkg/tenant/db"
-	ulids "github.com/rotationalio/ensign/pkg/utils/ulid"
+	"github.com/rotationalio/ensign/pkg/utils/ulids"
 	"github.com/stretchr/testify/require"
 )
