+++ pkg/utils/ulids/ulids.go
@@ -4,7 +4,7 @@ provides some common functionality (like checking if a ULID is null or is zero) 
 as a process-global, cryptographically random, monotonic, and thread-safe ulid
 generation mechanism that can be used from external packages.
 */
-package ulid
+package ulids
 
 import (
 	"crypto/rand"
